add edits kotlin part 1
