+++ app/src/main/java/com/almaz/task1/kotlin_training/part_1/Book.kt
@@ -1,24 +1,23 @@
 package com.almaz.task1.kotlin_training.part_1
 
-const val FLASH_FICTION_WORDS_COUNT = 1000
-const val SHORT_STORY_WORDS_COUNT = 7500
+const val FLASH_FICTION_WORD_COUNT_UPPER_BOUND = 1000
+const val SHORT_STORY_WORD_COUNT_LOWER_BOUND = 1001
+const val SHORT_STORY_WORD_COUNT_UPPER_BOUND = 7500
 const val FLASH_FICTION = "Flash Fiction"
 const val SHORT_STORY = "Short Story"
 const val NOVEL = "Novel"
 
 class Book(
-    override var price: Long,
+    override var price: Double,
     override var wordCount: Long
 ) : Publication {
 
     override fun getType(): String {
-        return when {
-            wordCount <= FLASH_FICTION_WORDS_COUNT -> FLASH_FICTION
-            wordCount <= SHORT_STORY_WORDS_COUNT -> SHORT_STORY
-            wordCount > SHORT_STORY_WORDS_COUNT -> NOVEL
-            else -> {
-                ""
-            }
+        return when (wordCount) {
+            in Long.MIN_VALUE..-1 -> throw Exception("Words count is negative!")
+            in 0..FLASH_FICTION_WORD_COUNT_UPPER_BOUND -> FLASH_FICTION
+            in SHORT_STORY_WORD_COUNT_LOWER_BOUND..SHORT_STORY_WORD_COUNT_UPPER_BOUND -> SHORT_STORY
+            else -> NOVEL
         }
     }
