+++ app/src/main/java/com/almaz/task1/kotlin_training/part_1/Magazine.kt
@@ -1,7 +1,9 @@
 package com.almaz.task1.kotlin_training.part_1
 
+const val MAGAZINE = "Magazine"
+
 class Magazine(
-    override val price: Long,
+    override val price: Double,
     override val wordCount: Long
 ) : Publication {
